only works with pysam 0.9.0
