+++ requirements.txt
@@ -3,5 +3,5 @@ pylint>=1.4.4
 pyensembl>=1.0.3
 varcode>=0.5.9
 pandas>=0.23.0
-pysam>=0.9.0
+pysam==0.9.0
 nose>=1.3.3
